ov017: promote daObjKsWater_c into a single C++ TU
The sunken ship's rising water (profile KS_MIZU) goes from seven
one-function files to one translation unit, src/actors/d_a_obj_ks_water.cpp,
that the ROM build compiles and byte-checks as a whole.

The class keeps the cartridge's own name. _ZTS at ov017:0x02111bc0 spells
"14daObjKsWater_c"; _ZTI at 0x02111bb4 is an __si_class_type_info whose +8
word points at _ZTI10dBgActor_c (ov002:0x021089ec), so the direct base is
dBgActor_c and nothing else; _ZTV at 0x02111bf8 is the address point, with a
zero offset-to-top and &_ZTI in the two words below it. The coined ShipWater
spelling is renamed out of the tree, and its duplicate _ZTV9ShipWater line --
which sat on the same word as _ZTV14daObjKsWater_c -- is dropped from
config/arm9/overlays/ov017/symbols.txt.

Six of the 34 vtable slots point into ov017: InitResources (0), Cleanup-
Resources (3), Behavior (6), Render (9) and the destructor pair (16, 17).
Everything else still holds dBgActor_c's arm9 word, so nothing else is
overridden. The four fields the class adds -- a TextureTransformer at 0x320,
mOriginalPosY, mChestsOpen and mSoundID -- account for 0x320..0x33f exactly,
which is the 0x340 the factory allocates.

The destructor is declared FIRST and INLINE. That is what makes this TU the
vtable's home; out of line, mwccarm emits D0 ahead of D1 while the cartridge
has D1 first, which objisolate refuses, and it also emits a D2 the ROM never
carried. Because this TU emits the vtable rather than importing it, the
factory's vptr store is &_ZTV14daObjKsWater_c[2] -- the symbol names the
emitted object's two-word preamble at 0x02111bf0, and [2] steps onto the
0x02111bf8 the cartridge writes.

The eleven RTTI records mwccarm emits as a consequence -- this class's trio
plus the dBase_c/fBase_c/dActor_c/dBgActor_c chain -- are licensed as
deadstrip-data against their canonical ROM addresses. romdata_check reports
6 VERIFIED and 5 PARTIAL for the file, 0 DIFFERS.

Proved with tools/rombuild.py -j16: 106/106 exact, 100.000000%,
mismatching 0.
