+++ attribution.json
@@ -1027,6 +1027,13 @@
     "src/actors/d_a_obj_km3_dorifu.cpp#_ZN17daObjKm3_Dorifu_c16CleanupResourcesEv": "github-actions[bot]",
     "src/actors/d_a_obj_km3_dorifu.cpp#_ZN17daObjKm3_Dorifu_cD0Ev": "github-actions[bot]",
     "src/actors/d_a_obj_km3_dorifu.cpp#_ZN17daObjKm3_Dorifu_cD1Ev": "github-actions[bot]",
+    "src/actors/d_a_obj_ks_water.cpp#_ZN14daObjKsWater_c13InitResourcesEv": "github-actions[bot]",
+    "src/actors/d_a_obj_ks_water.cpp#_ZN14daObjKsWater_c16CleanupResourcesEv": "github-actions[bot]",
+    "src/actors/d_a_obj_ks_water.cpp#_ZN14daObjKsWater_c6RenderEv": "github-actions[bot]",
+    "src/actors/d_a_obj_ks_water.cpp#_ZN14daObjKsWater_c8BehaviorEv": "github-actions[bot]",
+    "src/actors/d_a_obj_ks_water.cpp#_ZN14daObjKsWater_cD0Ev": "github-actions[bot]",
+    "src/actors/d_a_obj_ks_water.cpp#_ZN14daObjKsWater_cD1Ev": "github-actions[bot]",
+    "src/actors/d_a_obj_ks_water.cpp#daObjKsWater_c_classInit": "github-actions[bot]",
     "src/actors/d_a_obj_kumo.cpp#_ZN11daObjKumo_c13InitResourcesEv": "github-actions[bot]",
     "src/actors/d_a_obj_kumo.cpp#_ZN11daObjKumo_c16CleanupResourcesEv": "github-actions[bot]",
     "src/actors/d_a_obj_kumo.cpp#_ZN11daObjKumo_c6RenderEv": "github-actions[bot]",
