+++ config/converted-baseline.json
@@ -7,7 +7,7 @@
     "no_codegen_trick",
     "no_mangled_refs"
   ],
-  "count": 2596,
+  "count": 2597,
   "converted": [
     "src/ARMMathLoadState.c",
     "src/ARMMathSaveState.c",
@@ -1884,9 +1884,6 @@
     "src/_ZN9SeesawBob6RenderEv.cpp",
     "src/_ZN9SeesawBobD0Ev.cpp",
     "src/_ZN9SeesawBobD1Ev.cpp",
-    "src/_ZN9ShipWater16CleanupResourcesEv.cpp",
-    "src/_ZN9ShipWaterD0Ev.cpp",
-    "src/_ZN9ShipWaterD1Ev.cpp",
     "src/_ZN9SolidHeap10VGetNodeIDEv.cpp",
     "src/_ZN9SolidHeap10VSetNodeIDEj.cpp",
     "src/_ZN9SolidHeap11VDeallocateEPv.cpp",
@@ -2369,6 +2366,10 @@
     "src/actors/d_a_obj_km3_dorifu.cpp#_ZN17daObjKm3_Dorifu_c16CleanupResourcesEv",
     "src/actors/d_a_obj_km3_dorifu.cpp#_ZN17daObjKm3_Dorifu_cD0Ev",
     "src/actors/d_a_obj_km3_dorifu.cpp#_ZN17daObjKm3_Dorifu_cD1Ev",
+    "src/actors/d_a_obj_ks_water.cpp#_ZN14daObjKsWater_c16CleanupResourcesEv",
+    "src/actors/d_a_obj_ks_water.cpp#_ZN14daObjKsWater_c6RenderEv",
+    "src/actors/d_a_obj_ks_water.cpp#_ZN14daObjKsWater_cD0Ev",
+    "src/actors/d_a_obj_ks_water.cpp#_ZN14daObjKsWater_cD1Ev",
     "src/actors/d_a_obj_kumo.cpp#_ZN11daObjKumo_c6RenderEv",
     "src/actors/d_a_obj_kumo.cpp#_ZN11daObjKumo_c8BehaviorEv",
     "src/actors/d_a_obj_kumo.cpp#_ZN11daObjKumo_cD0Ev",
