+++ config/tu_manifest.d/ov098/ArrowSignRight.json
@@ -100,7 +100,7 @@
   "bss": [],
   "notes": [
     "Generated by tools/tubuild.py create. See the file's own header comment for what was and was not reconciled, and notes/translation-unit-reconstruction-plan.md section 7.3 for what this generator does and does not attempt to resolve automatically.",
-    "Second key-function-TU pilot after ov017/ShipWater, same shape as PoleLift/ShipWater: compiler_only_output covers D2Ev (byte-identical to D1, no ROM home, no inbound reloc). externalized_output covers the eight inherited vague-linkage RTTI records shared with the fBase_c/dBase_c/dActor_c/dBgActor_c hierarchy (arm9/ov002 canonical addresses, identical structure to ShipWater's own -- confirmed via a live ELF read of this TU's compiled object, not assumed). ArrowSignRight's own new _ZTI14ArrowSignRight/_ZTS14ArrowSignRight/_ZTV14ArrowSignRight stay unlicensed -- same as every prior key-function-TU pilot -- because no policy mechanism licenses a class's first-emitted vtable yet (Phase F, plan sec 15). NEW CASE NOT SEEN IN EITHER PRIOR PILOT: _ZN7Vector3D1Ev is also unlicensed. It is types.h's inline `~Vector3(){}`, re-emitted by every TU that destroys a Vector3, and it DOES have a configured ROM home (arm9 0x020072c0) -- confirmed identical body (`bx lr`) and zero inbound relocations within this TU via the same ELF read. Neither policy schema fits it: compiler_only_output refuses any symbol with a configured ROM home (it is not compiler-only), and externalized_output is restricted to _ZTI/_ZTS objects by its own validation, not .text functions. Left unlicensed alongside this TU's own vtable/RTTI; widening one of the two policy schemas (or adding a third) to cover a compiler-duplicated ordinary function is follow-up work, not attempted here. Render()/CleanupResources() were reconciled against the real header before verifying: the generator's local shadow struct pair (Base/Derived, standing in for the real inherited Model member) is gone in favour of `mModel.Render(0)`, matching ShipWater's precedent; and CleanupResources' mMeshCollider access, which the generator wrote as a raw offset cast to `dBgW*`, is now the plain member calls `mMeshCollider.IsEnabled()`/`.Disable()` -- the raw-cast form was ALSO a latent correctness bug once the file-table externs were retyped to `ArrowSignFileColumn`, not just a readability issue: the legacy file's byte-offset pointer arithmetic assumed a `char[]` base type, and multiplying an already-struct-typed pointer by 0xc again overshot every row lookup, which is what verify's first pass caught as a 999-word DIFF."
+    "Second key-function-TU pilot after ov017/daObjKsWater_c, same shape as PoleLift/daObjKsWater_c: compiler_only_output covers D2Ev (byte-identical to D1, no ROM home, no inbound reloc). externalized_output covers the eight inherited vague-linkage RTTI records shared with the fBase_c/dBase_c/dActor_c/dBgActor_c hierarchy (arm9/ov002 canonical addresses, identical structure to daObjKsWater_c's own -- confirmed via a live ELF read of this TU's compiled object, not assumed). ArrowSignRight's own new _ZTI14ArrowSignRight/_ZTS14ArrowSignRight/_ZTV14ArrowSignRight stay unlicensed -- same as every prior key-function-TU pilot -- because no policy mechanism licenses a class's first-emitted vtable yet (Phase F, plan sec 15). NEW CASE NOT SEEN IN EITHER PRIOR PILOT: _ZN7Vector3D1Ev is also unlicensed. It is types.h's inline `~Vector3(){}`, re-emitted by every TU that destroys a Vector3, and it DOES have a configured ROM home (arm9 0x020072c0) -- confirmed identical body (`bx lr`) and zero inbound relocations within this TU via the same ELF read. Neither policy schema fits it: compiler_only_output refuses any symbol with a configured ROM home (it is not compiler-only), and externalized_output is restricted to _ZTI/_ZTS objects by its own validation, not .text functions. Left unlicensed alongside this TU's own vtable/RTTI; widening one of the two policy schemas (or adding a third) to cover a compiler-duplicated ordinary function is follow-up work, not attempted here. Render()/CleanupResources() were reconciled against the real header before verifying: the generator's local shadow struct pair (Base/Derived, standing in for the real inherited Model member) is gone in favour of `mModel.Render(0)`, matching daObjKsWater_c's precedent; and CleanupResources' mMeshCollider access, which the generator wrote as a raw offset cast to `dBgW*`, is now the plain member calls `mMeshCollider.IsEnabled()`/`.Disable()` -- the raw-cast form was ALSO a latent correctness bug once the file-table externs were retyped to `ArrowSignFileColumn`, not just a readability issue: the legacy file's byte-offset pointer arithmetic assumed a `char[]` base type, and multiplying an already-struct-typed pointer by 0xc again overshot every row lookup, which is what verify's first pass caught as a 999-word DIFF."
   ],
   "verification": {
     "round": "tools/tubuild.py verify -- text-only, whole shadow TU compiled once",
