--- include/ShipWater.h
@@ -1,93 +1,0 @@
-#ifndef SHIPWATER_H
-#define SHIPWATER_H
-
-#include "types.h"
-#include "dBgW_KcMbg.h"
-
-/* Derives from dBgActor_c: the destructor stores this class's vtable, then
- * dBgActor_c's -- inlined -- then destroys the dBgW_KcMbg at 0x124 and
- * the Model at 0xd4 before chaining to dActor_c. All three belong to dBgActor_c.
- * Everything this header used to restate below 0x31e was dActor_c's and
- * dBgActor_c's, and is inherited now.
- *
- * SIZE IS THE OBSERVED FIELD SPAN, rounded up. It guards this declaration; it
- * is not independent evidence about the ROM.
- */
-
-#ifdef __cplusplus
-
-#include "dBgActor_c.h"
-#include "TextureTransformer.h"
-
-struct ShipWater : dBgActor_c {
-    u8  pad_31e[0x2];
-    TextureTransformer mTextureTransformer;/* 0x320 */
-    s32 mOriginalPosY;                      /* 0x334 */
-    u8 mChestsOpen;                       /* 0x338 */
-    u8  pad_339[0x3];
-    s32 mSoundID;                      /* 0x33c */
-
-    /* --- vtable --- */
-    virtual ~ShipWater();
-
-    int Behavior();
-    int CleanupResources();
-    int InitResources();
-    int Render();
-};
-
-typedef char ShipWater_size_must_be_0x340[sizeof(ShipWater) == 0x340 ? 1 : -1];
-
-#else
-
-/* The C spelling of the same object, flat. Kept because the D0 file is a C
-   translation unit that reads these fields, and D0 is compiler-generated so it
-   can never be migrated. Same arrangement as include/ShadowModel.h. */
-struct ShipWater {
-    u8  pad_000[0x60];
-    s32 mPosY;            /* 0x060 */
-    u8  pad_064[0x2a];
-    /* 0x08e..0x0d4 is dActor_c's, and dActor_c.h is de-bannered -- hand-reconstructed, not generated. Was one u8
-       marker over the whole range. */
-    s16 mAngleY;                 /* 0x08e */
-    s16 mAngleZ;                 /* 0x090 */
-    s16 mPrevAngleX;             /* 0x092 */
-    s16 mPrevAngleY;             /* 0x094 */
-    s16 mPrevAngleZ;             /* 0x096 */
-    s32 mHorzSpeed;              /* 0x098 */
-    s32 mVertAccel;              /* 0x09c */
-    s32 mTerminalVelocity;       /* 0x0a0 */
-    u8  pad_0a4[0x4];
-    s32 mVertSpeed;              /* 0x0a8 */
-    u8  pad_0ac[0x4];
-    u32 mFlags;                  /* 0x0b0 */
-    s32 mClipOffsetY;            /* 0x0b4 */
-    s32 mClipRadius;             /* 0x0b8 */
-    s32 mClipDistance;           /* 0x0bc */
-    s32 mFarDistance;            /* 0x0c0 */
-    u8  mClipResult;             /* 0x0c4 */
-    u8  pad_0c5[0x7];
-    s8  mAreaId;                 /* 0x0cc */
-    u8  pad_0cd[0x1];
-    s16 mDeathTableID;           /* 0x0ce */
-    u8  pad_0d0[0x4];
-    /* Model member, named by _ZN5ModelD1Ev at +0xd4 -- a relocation the ROM build checks.
-       D1 and not D2, so it is this type and not an inlined base. The marker's pad stopped
-       short of the object, so the member also takes over unk_0dc (+0x8 = data), which the
-       header declared separately inside it. */
-    Model mModel;            /* 0x0d4 */
-    /* dBgW_KcMbg member. The cartridge's own ~ShipWater calls _ZN10dBgW_KcMbgD1Ev at
-       +0x124 (D0/D1), a relocation the ROM build checks; recovered by
-       tools/dtor_members.py. D1 and not D2, so it is this type and not an inlined base. */
-    dBgW_KcMbg mMeshCollider;            /* 0x124 */
-    u8  pad_2ec[0x34];
-    TextureTransformer mTextureTransformer; /* 0x320 */
-    s32 mOriginalPosY;            /* 0x334 */
-    u8  mChestsOpen;            /* 0x338 */
-    u8  pad_339[0x3];
-    s32 mSoundID;            /* 0x33c */
-};
-
-#endif /* __cplusplus */
-
-#endif /* SHIPWATER_H */
