--- src_tu/actors/ShipWater.cpp
@@ -1,212 +1,0 @@
-//cpp
-/* SHADOW translation unit -- generated by tools/tubuild.py create.
- * ov017/ShipWater  (7 function(s))
- *
- * NOT ENROLLED, NOT CANONICAL. This file contributes nothing to the ROM
- * build. Render/CleanupResources/Behavior are reconciled against the real
- * headers (mModel/mMeshCollider/mTextureTransformer are real inherited
- * members; the legacy generator's local shadow structs and offset casts are
- * gone), the same reconciliation pilot #1 did by hand for PoleLift's
- * equivalents (see notes/tu-reconstruction-pilot-report.md sec 5.2). What is
- * left as extern "C" is load-bearing: Sound::PlayLong and the two Fix12<int>
- * SetFile calls have no real-method route yet (see the comments at each).
- *
- * FUNCTION ORDER IS DELIBERATELY THE REVERSE OF THE ROM'S -- mwccarm 2004/b56
- * emits one .text section per function, in the REVERSE of source order, so
- * the highest-address ROM function is written FIRST here. Do not reorder;
- * see notes/tu-reconstruction-pilot-report.md sec 3 for the one documented
- * exception (a destructor's D0/D1/D2 group has compiler-chosen order).
- *
- * Assembled from these legacy one-function sources (ROM address order):
- *   [0] 0x021111a0  src/_ZN9ShipWaterD1Ev.cpp
- *   [1] 0x021111ec  src/_ZN9ShipWaterD0Ev.cpp
- *   [2] 0x0211124c  src/_ZN9ShipWater16CleanupResourcesEv.cpp
- *   [3] 0x02111290  src/_ZN9ShipWater6RenderEv.cpp
- *   [4] 0x021112c4  src/_ZN9ShipWater8BehaviorEv.cpp
- *   [5] 0x021113c0  src/_ZN9ShipWater13InitResourcesEv.cpp
- *   [6] 0x02111480  src/daObjKsWater_c_classInit.c
- */
-
-/* Includes: union of the legacy files', first-seen in ROM-ascending
- * processing order. NOT verified for header ordering constraints (e.g. a
- * common.h-before-X rule) -- watch for new compile errors after this. */
-#include "ShipWater.h"
-#include "SharedFilePtr.h"
-#include "dBgW.h"
-#include "Model.h"
-#include "decl_ActorBase.h"
-#include "decl_Platform.h"
-#include "decl_TextureTransformer.h"
-
-/* Sound::PlayLong has no header declaration yet (Sound.h keeps only PlayBank3
- * and Player::SetPlayableSeqCount so far), and the position it takes by
- * const-reference is dActor_c's mCamSpacePosX/Y/Z -- three separate s32
- * fields, not a Vector3 -- so a real call still needs a reinterpret at the
- * call site either way. Left as the mangled free function until one of
- * those two gets fixed; not a shadow-struct workaround. */
-extern "C" {
-extern int _ZN5Sound8PlayLongEjjjRK7Vector3s(unsigned int a, unsigned int b, unsigned int cc, void* v, unsigned int e);
-extern int _ZTV9ShipWater[];
-extern SharedFilePtr data_ov017_02111c80;   /* the KCL handle */
-extern SharedFilePtr data_ov017_02111c88;   /* the BMD handle */
-extern BTA_File data_ov017_02111a60;
-extern CLPS_Block data_ov017_02111a94;
-void _ZN18TextureTransformer7SetFileER8BTA_Filei5Fix12IiEj(
-TextureTransformer *self, BTA_File *file, int flags, int speed, u32 startFrame);
-void _ZN10dBgW_KcMbg7SetFileEP8KCL_FileRK9Matrix4x35Fix12IiEsR10CLPS_Block(
-dBgW_KcMbg *self, KCL_File *file, const Matrix4x3 *mat, int scale, s16 angY,
-CLPS_Block *clps);
-}
-
-/* -------------------------------------------------------------------------- */
-/* ROM ordinal 6 -- daObjKsWater_c_classInit, 0x02111480, size 0x38 */
-/* -------------------------------------------------------------------------- */
-// @symbol daObjKsWater_c_classInit
-/* recovered: vtable identified, globals resolved, declarations from a shared header */
-/* recovered: vtable identified, globals resolved */
-/* resolved: VT0 = _ZTV9ShipWater */
-extern "C" {  /* .c-derived member: C linkage for the whole block */
-/* Reconstructed source-style name. Historical alias: ShipWater_Spawn. */
-int *daObjKsWater_c_classInit(void)
-{
-    int *p = (int *)_ZN7fBase_cnwEj(832);
-    if (p) {
-        _ZN10dBgActor_cC2Ev(p);
-        p[0] = (int)(_ZTV9ShipWater + 2);
-        _ZN18TextureTransformerC1Ev((char *)p + 0x320);
-    }
-    return p;
-}
-}
-
-/* -------------------------------------------------------------------------- */
-/* ROM ordinal 5 -- _ZN9ShipWater13InitResourcesEv, 0x021113c0, size 0xc0 */
-/* -------------------------------------------------------------------------- */
-// @symbol _ZN9ShipWater13InitResourcesEv
-/* ShipWater::InitResources -- load the water model and its BTA texture
- * animation, load and place the mesh collider, and remember the starting
- * surface height so Behavior can raise the water from it.
- *
- * THE COLLISION FILE'S HANDLE WAS NAMED FROM THE WRONG OVERLAY. The C form
- * passed `_ZN18TiltingPlatformBfsD0Ev', an ov045 FUNCTION that happens to sit
- * at 0x02111c80 -- the same address as ov017's own `data_ov017_02111c80', which
- * is what this actually is. Overlays share address space, so both names
- * relocate to the same word and the byte gate could not tell them apart (see
- * "Reloc slots are wildcards" in notes/mwccarm-codegen.md). Named correctly here.
- *
- * The two SetFile calls stay extern-C mangled free functions -- the same
- * arrangement LavaSeesaw and Scuttlebug use, for the same reason: both take
- * Fix12<int> by value, and Fix12<int> is an aggregate with no converting
- * constructor from the int literals these calls pass.
- *
- * The `+ 4' on the model handle is SharedFilePtr's loaded-file pointer.
- * include/SharedFilePtr.h declares no fields on purpose -- the layout is not
- * recovered and the files that spell it locally disagree -- so the offset stays
- * raw and says what it is rather than inventing a member.
- */
-int ShipWater::InitResources()
-{
-    mModel.SetFile((BMD_File *)Model::LoadFile(data_ov017_02111c88), 1, -1);
-    TextureTransformer::Prepare(**(BMD_File **)((char *)&data_ov017_02111c88 + 4),
-                                data_ov017_02111a60);
-    _ZN18TextureTransformer7SetFileER8BTA_Filei5Fix12IiEj(
-        &mTextureTransformer, &data_ov017_02111a60, 0, 0x1000, 0);
-    UpdateModelPosAndRotY();
-    UpdateClsnPosAndRot();
-    _ZN10dBgW_KcMbg7SetFileEP8KCL_FileRK9Matrix4x35Fix12IiEsR10CLPS_Block(
-        &mMeshCollider, (KCL_File *)dBgW_Kc::LoadFile(data_ov017_02111c80),
-        &mClsnMat, 0x1000, mAngleY, &data_ov017_02111a94);
-    mMeshCollider.Enable(this);
-    mOriginalPosY = mPosY;
-    return 1;
-}
-
-/* -------------------------------------------------------------------------- */
-/* ROM ordinal 4 -- _ZN9ShipWater8BehaviorEv, 0x021112c4, size 0xfc */
-/* -------------------------------------------------------------------------- */
-// @symbol _ZN9ShipWater8BehaviorEv
-/* recovered: named members + shared header, real C++ method */
-int ShipWater::Behavior()
-{
-    if (mChestsOpen == 0) {
-        int ok = 1;
-        dActor_c* p = dActor_c::FindWithActorID(0xd, 0);
-        while (p != 0) {
-            /* offset into the chest actor's own (unrecovered) type -- ID 0xd is not
-             * ShipWater or any class this file's headers know about */
-            int state = *(int*)((char*)p + 0x16c);
-            int cond = 1;
-            if (state != 1) {
-                if (state != 2) cond = 0;
-            }
-            if (cond == 0) ok = 0;
-            p = dActor_c::FindWithActorID(0xd, p);
-        }
-        if (ok != 0) mChestsOpen = 1;
-    } else {
-        int d = mOriginalPosY - mPosY;
-        if (d < 0) d = -d;
-        if (d < 0x92e000) {
-            int* q;
-            mSoundID = _ZN5Sound8PlayLongEjjjRK7Vector3s(mSoundID, 3, 0x96, ((char*)this)+0x74, 0);
-            q = (int*)((char*)&mPosY);
-            *q -= 0x5000;
-        }
-    }
-    mTextureTransformer.speed = 0x1000;
-    mTextureTransformer.Advance();
-    UpdateModelPosAndRotY();
-    UpdateClsnPosAndRot();
-    return 1;
-}
-
-/* -------------------------------------------------------------------------- */
-/* ROM ordinal 3 -- _ZN9ShipWater6RenderEv, 0x02111290, size 0x34 */
-/* -------------------------------------------------------------------------- */
-// @symbol _ZN9ShipWater6RenderEv
-/* recovered: named members + shared header, real C++ method */
-int ShipWater::Render()
-{
-    mTextureTransformer.Update(mModel.data);
-    mModel.Render(0);
-    return 1;
-}
-
-/* -------------------------------------------------------------------------- */
-/* ROM ordinal 2 -- _ZN9ShipWater16CleanupResourcesEv, 0x0211124c, size 0x44 */
-/* -------------------------------------------------------------------------- */
-// @symbol _ZN9ShipWater16CleanupResourcesEv
-/* recovered: named members + shared header, real C++ method, declarations from a shared header */
-/* recovered: named members + shared header, real C++ method */
-int ShipWater::CleanupResources()
-{
-    if (mMeshCollider.IsEnabled()) {
-        mMeshCollider.Disable();
-    }
-    data_ov017_02111c88.Release();
-    data_ov017_02111c80.Release();
-    return 1;
-}
-
-/* -------------------------------------------------------------------------- */
-/* ROM ordinal 1 -- _ZN9ShipWaterD0Ev, 0x021111ec, size 0x60 */
-/* ROM ordinal 0 -- _ZN9ShipWaterD1Ev, 0x021111a0, size 0x4c */
-/* -------------------------------------------------------------------------- */
-// @symbol _ZN9ShipWaterD0Ev
-// @symbol _ZN9ShipWaterD1Ev
-/* recovered: real C++ destructor -- the compiler emits D2, D0, and D1 together
- * from this ONE source definition (decl_common's tu-reconstruction collapse
- * rule); the two legacy per-variant files were merging the same body twice,
- * which mwcc rejects as a redefinition. objisolate keeps each emitted
- * .text section bound to its own symbol.
- *
- * Two vtable stores and three destructor calls, every one a consequence of
- * `struct ShipWater : dBgActor_c`: its own vptr, then dBgActor_c's -- inlined,
- * because dBgActor_c's destructor is defined in its class body -- then
- * dBgActor_c's Model and dBgW_KcMbg, then dActor_c. This class adds no
- * member with a destructor of its own. D0 additionally destroys through the
- * base and returns the object to its heap via an inline operator delete.
- */
-ShipWater::~ShipWater()
-{
-}
