+++ symbols/profile_reconstruction_registry.json
@@ -2910,7 +2910,7 @@
       "overlay_ambiguous": false,
       "overlay": "ov017",
       "profile_address": "0x02111bd4",
-      "current_profile_name": "ShipWater_SpawnInfo",
+      "current_profile_name": "daObjKsWater_c_SpawnInfo",
       "proposed_profile_name": "g_profile_KS_MIZU",
       "descriptor_layout": "actor_profile_0x1c",
       "profile_index": 62,
@@ -2926,8 +2926,8 @@
       "far_distance": "0x00000000",
       "factory_module": "ov017",
       "factory_address": "0x02111480",
-      "current_factory_name": "ShipWater_Spawn",
-      "current_factory_file": "src/ShipWater_Spawn.c",
+      "current_factory_name": "daObjKsWater_c_Spawn",
+      "current_factory_file": "src/daObjKsWater_c_Spawn.c",
       "proposed_factory_name": "daObjKsWater_c_classInit",
       "factory_shape": "new_plus_inlined_ctor",
       "alloc_size": "0x00000340",
