+++ symbols/profile_reconstruction_registry.tsv
@@ -62,7 +62,7 @@ actor_id	profile_id	profile_id_source	debug_string_address	candidate_ordinal	reg
 59	KI_HASIRA_DAI	symbols/actor_debug_names.tsv (literal ROM string table)	0x020905c8	1	1	False	ov102	0x0214e0fc	RockTriangle_SpawnInfo	g_profile_KI_HASIRA_DAI	actor_profile_0x1c	59	True	59	183		unlocated: no separate group-flags field is read from this descriptor by the recovered fBase_c/dActor_c constructors	0x00000002	0x00000000	0x00200000	0x02000000	0x00000000	ov102	0x02148f5c	RockTriangle_Spawn	src/RockTriangle_Spawn.c	daObjSimpleBg_c_classInit	new_plus_inlined_ctor	0x00000320	0x0214e1d8	daObjSimpleBg_c	SM64DS RTTI ov102:0x0214e0c0	ov102:0x02148e9c;ov102:0x02148ecc;ov102:0x02148efc;ov102:0x02148f2c;ov102:0x02148f5c;ov102:0x02148f8c	d_a_obj_simple_bg.cpp	0x2148a80	0x2148fbc	yes	no	not_apply=global_classinit_name_collision
 60	KI_ITA	symbols/actor_debug_names.tsv (literal ROM string table)	0x0208f7e8	1	1	False	ov016	0x02114ba8	FloatOnWaterPlatformJrb_SpawnInfo	g_profile_KI_ITA	actor_profile_0x1c	60	True	60	184		unlocated: no separate group-flags field is read from this descriptor by the recovered fBase_c/dActor_c constructors	0x00000002	0x00000000	0x00250000	0x02000000	0x00000000	ov016	0x02112fbc	FloatOnWaterPlatformJrb_Spawn	src/FloatOnWaterPlatformJrb_Spawn.c	daObjKi_Ita_c_classInit	new_plus_inlined_ctor	0x00000348	0x02114bcc	daObjKi_Ita_c	SM64DS RTTI ov016:0x02114b80		d_a_obj_ki_ita.cpp	0x2112ef4	0x2112ff8	yes	yes	eligible=unique_registry_context+rom_profile_id+unique_class_factory
 61	KI_IWA	symbols/actor_debug_names.tsv (literal ROM string table)	0x0208f7a0	1	1	False	ov102	0x0214e0e0	StaticRock_SpawnInfo	g_profile_KI_IWA	actor_profile_0x1c	61	True	61	185		unlocated: no separate group-flags field is read from this descriptor by the recovered fBase_c/dActor_c constructors	0x00000002	0x00000000	0x00500000	0x02000000	0x00000000	ov102	0x02148f8c	StaticRock_Spawn	src/StaticRock_Spawn.c	daObjSimpleBg_c_classInit	new_plus_inlined_ctor	0x00000320	0x0214e1d8	daObjSimpleBg_c	SM64DS RTTI ov102:0x0214e0c0	ov102:0x02148e9c;ov102:0x02148ecc;ov102:0x02148efc;ov102:0x02148f2c;ov102:0x02148f5c;ov102:0x02148f8c	d_a_obj_simple_bg.cpp	0x2148a80	0x2148fbc	yes	no	not_apply=global_classinit_name_collision
-62	KS_MIZU	symbols/actor_debug_names.tsv (literal ROM string table)	0x0208f8d0	1	1	False	ov017	0x02111bd4	ShipWater_SpawnInfo	g_profile_KS_MIZU	actor_profile_0x1c	62	True	62	186		unlocated: no separate group-flags field is read from this descriptor by the recovered fBase_c/dActor_c constructors	0x00000000	0x00000000	0x0000a000	0x00000000	0x00000000	ov017	0x02111480	ShipWater_Spawn	src/ShipWater_Spawn.c	daObjKsWater_c_classInit	new_plus_inlined_ctor	0x00000340	0x02111bf8	daObjKsWater_c	SM64DS RTTI ov017:0x02111bb4		d_a_obj_ks_water.cpp	0x21111a0	0x21114b8	yes	yes	eligible=unique_registry_context+rom_profile_id+unique_class_factory
+62	KS_MIZU	symbols/actor_debug_names.tsv (literal ROM string table)	0x0208f8d0	1	1	False	ov017	0x02111bd4	daObjKsWater_c_SpawnInfo	g_profile_KS_MIZU	actor_profile_0x1c	62	True	62	186		unlocated: no separate group-flags field is read from this descriptor by the recovered fBase_c/dActor_c constructors	0x00000000	0x00000000	0x0000a000	0x00000000	0x00000000	ov017	0x02111480	daObjKsWater_c_Spawn	src/daObjKsWater_c_Spawn.c	daObjKsWater_c_classInit	new_plus_inlined_ctor	0x00000340	0x02111bf8	daObjKsWater_c	SM64DS RTTI ov017:0x02111bb4		d_a_obj_ks_water.cpp	0x21111a0	0x21114b8	yes	yes	eligible=unique_registry_context+rom_profile_id+unique_class_factory
 63	SM_LIFT	symbols/actor_debug_names.tsv (literal ROM string table)	0x0208f8e0	1	1	False	ov018	0x021138a8	SkiLift_SpawnInfo	g_profile_SM_LIFT	actor_profile_0x1c	63	True	63	230		unlocated: no separate group-flags field is read from this descriptor by the recovered fBase_c/dActor_c constructors	0x00000002	0x00000000	0x00300000	0x01000000	0x00000000	ov018	0x02111818	SkiLift_Spawn	src/SkiLift_Spawn.c	daObjSm_Lift_c_classInit	new_plus_inlined_ctor	0x00000354	0x021138cc	daObjSm_Lift_c	SM64DS RTTI ov018:0x02113888		d_a_obj_sm_lift.cpp	0x21111a0	0x2111848	yes	yes	eligible=unique_registry_context+rom_profile_id+unique_class_factory
 64	CASKET	symbols/actor_debug_names.tsv (literal ROM string table)	0x0208f800	1	1	False	ov071	0x02122eb0	Coffin_SpawnInfo	g_profile_CASKET	actor_profile_0x1c	64	True	64	159		unlocated: no separate group-flags field is read from this descriptor by the recovered fBase_c/dActor_c constructors	0x02000002	0x00100000	0x00300000	0x01000000	0x00000000	ov071	0x02122670	Coffin_Spawn	src/Coffin_Spawn.c	daObjCasket_c_classInit	new_plus_inlined_ctor	0x0000032c	0x02122efc	daObjCasket_c	SM64DS RTTI ov071:0x02122e94		d_a_obj_casket.cpp	0x2121fe4	0x21226a0	yes	yes	eligible=unique_registry_context+rom_profile_id+unique_class_factory
 65	KAIDAN	symbols/actor_debug_names.tsv (literal ROM string table)	0x0208f808	1	1	False	ov063	0x0211ea10	MansionSteps_SpawnInfo	g_profile_KAIDAN	actor_profile_0x1c	65	True	65	53		unlocated: no separate group-flags field is read from this descriptor by the recovered fBase_c/dActor_c constructors	0x00000000	0x00064000	0x00064000	0x01000000	0x00a00000	ov063	0x0211d360	MansionSteps_Spawn	src/actors/MansionSteps/MansionSteps_Spawn.c	daTrsTrap_c_classInit	new_plus_inlined_ctor	0x00000354	0x0211ea88	daTrsTrap_c	SM64DS RTTI ov063:0x0211e9cc	ov063:0x0211d2a0;ov063:0x0211d2e0;ov063:0x0211d320;ov063:0x0211d360	d_a_trs_trap.cpp	0x211c600	0x211d3a0	yes	no	not_apply=global_classinit_name_collision
